refactor(axon_rust-1d2.1.2): split build_config.rs below monolith limit
Move 9 helper functions from build_config.rs to helpers.rs:
- parse_csv_env, parse_origin_allowlist, read_env, default_sqlite_path
- validate_collection_name (with tests)
- validate_custom_headers
- resolve_mcp_transport
- resolve_ask_adapter_cmd, resolve_ask_adapter_args (with tests)
- env_port

Also adds env_bool_opt() to helpers.rs for use by 1d2.1.3 TOML wiring.
build_config.rs: 971 → 681 lines (in config/** — exempt from 500-line hard limit).
helpers.rs: 97 → 413 lines (passes monolith check).
